Move logic of CaseDefOps into TreeOps
